// frontend/src/generated/api.ts
export interface components {
  schemas: {
// ...
    Product: {
      /** Format: uuid */
      id: string;
      /** Format: uuid */
      siteAnalysisId: string;
      title: string;
      brand?: string;
      year?: string;
      category?: string;
      subCategory?: string;
      description?: string;
      commercialDescription?: string;
      /** Format: double */
      price?: number;
      /** @default EUR */
      currency: string;
      /** Format: uri */
      imageUrl?: string;
      /** Format: uri */
      sourceUrl?: string;
      /** Format: date-time */
      createdAt: string;
    };
    ProductListResponse: {
      products: components['schemas']['Product'][];
      total: number;
      brands: string[];
      categories: string[];
      subCategories: string[];
      years: string[];
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}